+++ resources/js/Components/Notifications/NotificationModal.vue
@@ -23,25 +23,25 @@ const notifications = ref([]);
 const closeModal = ref(null);
 const notificationsDialog = ref(null);
 
-const fetchNotifications = async () => {
-    // Make an API request to fetch notifications
-    const response = await fetch(`/notifications`);
-    const data = await response.json();
-    // userStore.notifications = data.notifications;
-    if (data.notifications && Array.isArray(data.notifications)) {
-        userStore.newNotifications = data.notifications.length;
-    } else {
-        // Handle the case where notifications are missing or not an array
-        userStore.newNotifications = 0; // or some other default value or error handling logic
-    }
-};
+// const fetchNotifications = async () => {
+//     // Make an API request to fetch notifications
+//     const response = await fetch(`/notifications`);
+//     const data = await response.json();
+//     // userStore.notifications = data.notifications;
+//     if (data.notifications && Array.isArray(data.notifications)) {
+//         userStore.newNotifications = data.notifications.length;
+//     } else {
+//         // Handle the case where notifications are missing or not an array
+//         userStore.newNotifications = 0; // or some other default value or error handling logic
+//     }
+// };
 
 function closeModalFunction() {
     closeModal.value.submit();
 }
 
 onMounted(() => {
-    fetchNotifications();
+    // fetchNotifications();
 });
 
 </script>